Refactor: move CurrentUserArgumentResolver
