+++ api-user/src/main/java/com/user/config/WebConfig.java
@@ -1,5 +1,6 @@
 package com.user.config;
 
+import com.user.config.security.CurrentUserArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
