+++ api-user/src/main/java/com/user/config/security/CurrentUserArgumentResolver.java
@@ -1,7 +1,5 @@
-package com.user.config;
+package com.user.config.security;
 
-import com.user.config.security.CurrentUser;
-import com.user.config.security.UserPrincipal;
 import com.user.utils.error.CommonException;
 import org.springframework.core.MethodParameter;
 import org.springframework.security.core.Authentication;
